+++ src/main/java/zaftnotameni/creatania/ponder/StringFunctions.kt
@@ -1,0 +1,7 @@
+package zaftnotameni.creatania.ponder
+
+import com.simibubi.create.foundation.fluid.FluidIngredient.FluidStackIngredient
+import net.minecraft.world.level.block.Block
+
+fun nameOf(fsi : FluidStackIngredient) : String = fsi?.getMatchingFluidStacks()?.first()?.displayName?.string ?: "UNKNOWN"
+fun nameOf(block : Block) : String = block?.name?.string  ?: "UNKNOWN"